+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.4.0] - 2025-06-23
 
 ### Added
 - Uniform live controls for `float`, `int` and `bool`
@@ -50,3 +50,4 @@ First version.
 [0.1.0]: https://github.com/ubitux/ShaderWorkshop/releases/tag/v0.1.0
 [0.2.0]: https://github.com/ubitux/ShaderWorkshop/compare/v0.1.0...v0.2.0
 [0.3.0]: https://github.com/ubitux/ShaderWorkshop/compare/v0.2.0...v0.3.0
+[0.4.0]: https://github.com/ubitux/ShaderWorkshop/compare/v0.3.0...v0.4.0